Testing and fixing errors
Application: fixed errors and made sure everything works.
-> only thing that doesn't work is that a user (that is not the first
node) cannot log in. (because front is set to null the whole time)

Did not do anything to database, other than delete a space.

Person: we need to implement fencing, so that there is no empty line before or after the text file.